refactor(listen): split extended player tabs
